Corrected termination of ortemanager
Sometimes I've realized that it was not possible to terminate ortemanager
by pressing Ctrl-C. It seems it was the case when I disconnecte from the
network and wanted to finish ortemanager later.

Since it is not safe to call pthread_mutex_lock() from signal handlers,
I've reworked termination to use sigwaitinfo() to wait for termination
signals.