JORTE: fix memory leak and add DomainEvents to DomainMgr
Memory leaks in jORTEDomainAppCreate() and jORTEDomainMgrCreate()
(domain events callback structure) are fixed. Domain events callbacks
are placed in a separate file. Their prototypes are placed in "jorte/
/jorte_protos_api.h". This way they could be used by DomainApp
and DomainMgr.

Native functions in both DomainApp.java and DomainMgr.java are now non-
static. This allows native domain destructor to free the memory allocated
to domain events callback structure.

It is now also possible to pass null for DomainEvents or DomainProp to
DomainApp and DomainMgr constructors.

The header file onLoad.h is removed and the findClass() prototype is placed
into "jorte/jorte_protos_api.h".