ORTE: add support for param structure in TypeRegister
This adds support for user defined callback function processEndianness()
with processParam structure being set at Type creation and passed
to processEndianness().

The processEndianness() function is called if defined and if there is no
deserialization function and thus no other way for the user application to
know the byte order of received issue.