Fix reported problem "compile error on cygwin"
The Windows service support in ORTE manager is enabled
now for Cygwin build as well. Build under MINGW has been
tested but some external volunteer is required for Cygwin
environment test.

Another option is to add config option which allows
to specify if Windows service support should or should not
be build.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>