Update versioning of shared libraries
Previously, we used -release to specify the SONAME, which would cause
binary incompatibility between releases even if the library is not
changed at all.  Additionally, libtool interface numbers were derived
from relase number in an undocumented way. This is removed as well and
now we have to modify SONAME manually, when the interface changes.