Java: Rewrite handling of ORTERecvInfo structure in middle-code
This commit should result in significant increase of subscriber's
performance in Java.

Previously we used to create many Java objects during the reception of fresh
issue. These objects carried the same information as the ORTERecvInfo C
structure. Now we handle it in a way similar to how data reception is taken
care of.

That is, we allocate a direct ByteBuffer upon the first invocation
of the C receive callback function. This buffer's address is then stored
in a helper structure, which is passed to this receive callback function.
When new issue arrives, we copy it's ORTERecvInfo C structure into this
ByteBuffer by calling memcpy(). We are not immediately parsing the contents
of this buffer for performance reasons. We do it only when someone
requests information included in the buffer (this is handled in the
org.ocera.orte.types.RecvInfo class).