JAVA: Speedup fresh issue reception and minor clean-up
This commit simplifies the receive callback. The process of getting pointer
to direct ByteBuffer holding copy of ORTERecvInfo struct is moved into the
jORTESubscriptionCreate() function, because the address remains the same
during life of the subscription. Method ID's of read() function of MessageData
class and callback() of ReceiveCallback class are cached in JORTECallbackContext_t
struct in jORTESubscriptionCreate() function, since the IDs remain valid during
the entire JVM session. Also many unnecessary checks are removed from the
receive callback. In overall this patch increases the probability, that the
callback function remains in cache, because many procedures are moved away.