shape_android: do not enforce ByteOrder
Setting of ByteOrder by user is not supported for publications. That is
on the Java part we discover, what the native ByteOrder is and set it
accordingly (see MessageData() constructor at
orte/java/src/org/ocera/orte/types/MessageData.java:41). On the C side it
is assumed, that the byte order is also native
(see ORTEPublicationSendLocked() at orte/liborte/ORTEPublication.c:316).

But if we set the ByteOrder of publisher different than native on the Java
part, than in liborte it is still believed to be native causing problems
during deserialization.