shape_android: fix freezes on app restart
When application was exited through 'back' button domain associated with
the running instance was destroyed, however the publishers and subscribers
registered within that domain were left intact.

This caused 'Application Not Responding' to be thrown when user wants to
start the application again, because it tries to use handles associated
with the now defunct publishers/subscribers.

In order to prevent this, we destroy Publisher and Subscriber objects upon
invocation of onDestroy() callback method.