SHAPE: Fix undefined behavior resulting in strange lock-ups
This commit fixes intermittent freezes of 'subscriber' dialog. These
freezes were caused by invoking QWidget::update() directly from a non-GUI
thread. This is undefined behavior and leads to many strange lock-ups and
even SEGFAULTS.

In order to fix this we now introduce the signal changed(), that will be
emitted if repaint of shapes is requested either from publisher or
subscriber. This signal is connected to the update() slot. By default
signals are connected to slots by Qt::AutoConnection connection type. This
should ensure, that QT decides internally if it should treat this
connection as Qt::DirectConnection in case of publisher and
Qt::QueuedConnection in case of subscriber. This decision is made when the
signal is emitted.